requirements_pavel.txt
cython
setuptools
numpy>=1.16.*
imgaug>=0.4.*
opencv-python==4.2.0.34
flask
flask-sqlalchemy
scikit_image
asyncio
gitpython
pycocotools
ujson
pillow==9.0.1
tqdm
matplotlib
scipy
torch>=1.8
torchvision>=0.9
PyYAML>=5.4
seaborn
pytorch_lightning
ipywidgets
gevent
termcolor
sklearn
aiogram

# git repos
PyTurboJPEG @ git+https://github.com/lilohuang/PyTurboJPEG.git
modelhub_client @ git+https://github.com/ria-com/modelhub-client.git